refactor: move data fetching logic to a separate helper function and clean up BlogPost component
